moved DependencyNode innver class out-of DependencyTreeParser class
